Move debug functions to utils directory.
